/ ‘Euler/8.hs’

+++ Euler/8.hs
@@ -76,5 +76,5 @@ isDivisibleBy n (f:fs) = ( n `mod` f == 0 )  && isDivisibleBy n fs
 
 main = do
   number_string <- getContents
-  let n = read number_string
+  let n::Int = read number_string
   putStrLn $ show n
